Split unit tests in two files
